axiom_factory/pyproject.toml
[project]
name = "axiom-station"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents-arcade>=0.0.4",
    "arcadepy>=1.8.0",
    "fastapi>=0.119.1",
    "firebase-admin>=7.1.0",
    "litellm>=1.78.5",
    "openai-agents>=0.4.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]